\name{Limpet Palettes}
\alias{Limpet Palettes}
\title{Limpet Palettes}
\usage{
lim_palette("pal1",n, c("discrete","continuous"))
}
\description{
Contains palettes based on limpet painting and can produce new continuous palettes following the design of the [wesanderson](https://github.com/karthik/wesanderson) package (see example).
}
\examples{
lim_palette("pal1",15,"continuous")
}
