\name{Limpet Palletes}
\alias{Limpet Palletes}
\title{Limpet Palletes}
\usage{
Lim_palette("pal1",n, c("discrete","continuous"))
}
\description{
Contains palettes based on limpet painting and can produce new continuous palettes following the design of the wesanderson package (see example).
}
\examples{
Lim_palette("pal1",15,"continuous")
}
